requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.57.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.2.13
kiwisolver==1.4.8
lxml==5.3.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
narwhals==1.36.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.51
Protego==0.4.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
queuelib==1.8.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rpds-py==0.24.0
scipy==1.15.2
Scrapy==2.12.0
scrapy-user-agents==0.1.1
selenium==4.31.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
tldextract==5.2.0
tornado==6.4.2
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typing_extensions==4.13.2
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.4.0
user-agents==2.2.0
w3lib==2.3.1
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.interface==7.2
